upload ready including db ref
