+++ react_postgreSQL_photo_album/myPhotos_api/controller/image.js
@@ -30,7 +30,7 @@ const uploadFromUrlHandler = async (req, res, db) => {
  };
 
 
-const handleImage = (req, res, db) => {
+const handleImage =(req, res, db) => {
   const {id} = req.body;
   db('users')
     .where('id' ,'=' , id)
