+++ react_postgreSQL_photo_album/myPhotos_api/db.js
@@ -1,0 +1,15 @@
+const knex = require('knex');
+require('dotenv').config()
+const {HOST, USER, DB_NAME} = process.env
+
+const db = knex({
+    client: 'pg',
+    connection: {
+      host : HOST,
+      user : USER,
+      password : 'password',
+      database : DB_NAME
+    }
+  });
+
+module.exports = db;
